pyproject.toml
[project]
name = "ai-emp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "langchain>=1.2.18",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.3",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.57.0",
    "tavily>=1.1.0",
    "uvicorn>=0.46.0",
]
